fix scripts

+++ dot-files/meow/.config/matugen/wallpaper.sh
@@ -1,7 +1,10 @@
 #!/usr/bin/env bash
 WALL="$1"
 
-rm ~/.cache/last-wallpaper
+
+if [[ -f ~/.cache/last-wallpaper ]]; then
+    rm ~/.cache/last-wallpaper
+fi
 cp "$WALL" ~/.cache/last-wallpaper
 
 matugen image "$WALL"
